adding hot module reload
