+++ .babelrc
@@ -10,6 +10,7 @@
     }]
   ],
   "plugins":[
+    "react-hot-loader/babel",
     "babel-plugin-transform-class-properties"
   ],
   "env":{
