+++ js/ClientApp.jsx
@@ -1,21 +1,13 @@
 import React from 'react';
 import { render } from 'react-dom';
-import { BrowserRouter, Route, Switch } from 'react-router-dom';
-import Landing from './Landing';
-import Search from './Search';
+import App from './app'
 
-const FourOhFour = () => <h1>404</h1>;
-
-const App = () => (
-  <BrowserRouter>
-    <div className="app">
-      <Switch>
-        <Route exact path="/" component={Landing} />
-        <Route exact path="/search" component={Search} />
-        <Route component={FourOhFour} />
-      </Switch>
-    </div>
-  </BrowserRouter>
-);
-
-render(<App />, document.getElementById('app'));
+const renderApp = () =>{
+  render(<App/>,document.getElementById('app'))
+}
+renderApp();
+if(module.hot){
+  module.hot.accept('./App', ()=>{
+    renderApp();
+  })
+}
