+++ js/Landing.jsx
@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom';
 
 const Landing = () => (
   <div className="landing">
-    <h1>videos</h1>
+    <h1>Nixalar Videos</h1>
     <input type="text" placeholder="search" />
     <Link to="/search">or Browse All</Link>
   </div>
